+++ src/app/shared/module/poe/component/item-frame-properties/item-frame-properties.component.ts
@@ -50,6 +50,7 @@ export class ItemFramePropertiesComponent implements OnInit {
       'prophecyText',
       'durability',
       'storedExperience',
+      'memoryStrands',
       'gemQualityType',
       'areaLevel',
       'ultimatum',