+++ src/app/shared/module/poe/type/item.type.ts
@@ -200,6 +200,7 @@ export interface ItemProperties {
   prophecyText?: string
   durability?: ItemValueProperty
   storedExperience?: ItemValueProperty
+  memoryStrands?: ItemValueProperty
   ultimatum?: ItemPropertiesUltimatum
   incursion?: ItemPropertiesIncursion
   heist?: ItemPropertiesHeist